+++ tests/testthat/test-readMaxQuantXML.R
@@ -9,18 +9,19 @@ test_that("reading MQ XML file works", {
                  "'<' not found")
 
     ## Read proper file and check that output is a
-    ## correctly named list of length 15
+    ## correctly named list of length 17
     mq <- readMaxQuantXML(
         system.file("extdata", "mq_example", "1356_mqpar.xml",
                     package = "einprot"))
     expect_type(mq, "list")
-    expect_equal(length(mq), 15)
+    expect_equal(length(mq), 17)
     expect_named(mq, c("MaxQuant version", "Parameter file", "Search engine",
                        "Raw file location", "Raw files", "Sample names",
                        "Databases", "Contaminants", "Quantification mode",
                        "Quantification settings (LFQ)", "Min. razor peptides",
                        "Requantify", "Enzymes", "Variable modifications",
-                       "Fixed modifications"))
+                       "Fixed modifications", "Max peptide mass",
+                       "Min peptide length"))
 
     ## All entries should be scalar values
     expect_true(all(vapply(mq, length, 0) == 1))
@@ -36,6 +37,8 @@ test_that("reading MQ XML file works", {
     expect_equal(mq$Enzymes, "Trypsin/P")
     expect_equal(mq$`Variable modifications`, "Oxidation (M), Acetyl (Protein N-term)")
     expect_equal(mq$`Fixed modifications`, "")
+    expect_equal(mq$`Max peptide mass`, "8000")
+    expect_equal(mq$`Min peptide length`, "7")
 
     ## Missing file
     mq <- readMaxQuantXML(mqParameterFile = NULL)