+++ NEWS.md
@@ -6,6 +6,7 @@
 * Remove features with too few non-imputed values before doing PCA
 * Retain a few more rowData columns for MaxQuant
 * Fix extraction of fixed modifications for MQ v2.2
+* Add max peptide mass and min peptide length to MQ report
 
 # einprot 0.7.0